added assert(fib(50) == 12586269025); and adjustet fib() return type to long long

+++ dynamic_programming_c/fibonacci.h
@@ -2,7 +2,7 @@
 #ifndef _FIBONACCI_
 #define _FIBONACCI_
 
-int fib(int n);
+long long fib(int n);
 
 #endif // !_FIBONACCI_
